move Channel.ts to ChannelHelper.ts
