+++ packages/utils/src/index.ts
@@ -1,4 +1,4 @@
 'use strict';
 
-export * from './helper/Channel.js';
+export * from './helper/ChannelHelper.js';
 export * from './helper/TypeCheck.js';
